End-stage Liver Disease Data

A dataset of patients evaluated for liver transplantation at the University Hospital of Leipzig from November 2012 to June 2015.

eldd

Format

A data.frame with 654 rows and 48 columns:

Age

integer, patient's age in years.

Sex

factor, patient's sex encoded as "M" for male and "W" for female.

DaysAtRisk

integer, number of days from the time when the laboratory measurements were taken to the last follow up.

Deceased

integer, was the patient alive (0) or deceased (1) on the last follow up.

LTx

integer, 1 if the patient got a liver transplantation else 0.

Cirrhosis

integer, 1 if the patient had a cirrhosis else 0.

ALF

integer, 1 if the patient had acute liver failure else 0.

Ethyltoxic

integer, 1 if the patient had a ethyltoxic liver disease else 0.

HBV

integer, 1 if the patient had a liver disease caused by an infection with the hepatitis B virus else 0.

HCV

integer, 1 if the patient had a liver disease caused by an infection with the hepatitis C virus else 0.

AIH

integer, 1 if the patient had an autoimmune hepatitis else 0.

PBC

integer, 1 if the patient had a primary sclerosing cirrhosis else 0.

PSC

integer, 1 if the patient had a primary biliary cirrhosis else 0.

NASH

integer, 1 if the patient had a non-alcoholic steatohepatitis else 0.

Cryptogenic

integer, 1 if the patient had cirrhosis of unknown origin else 0.

Dialysis

integer, 1 if the patient required renal replacement therapy else 0.

GIB

integer, 1 if the patient suffered gastrointestinal bleedings else 0.

SBP

integer, 1 if the patient suffered spontaneous bacterial peritonitis else 0.

HCC

integer, 1 if the patient suffered hepatocellular carcinoma else 0.

ALAT_S

numeric, serum level of ALAT, alanine aminotransferase in µkat/l.

ALB_S

numeric, serum level of albumin in g/l.

AP_S

numeric, serum level of alkaline phosphatase in µkat/l.

ASAT_S

numeric, serum level of ASAT, aspartate aminotransferase in µkat/l.

B_MPV_E

numeric, mean platelet volume in fl.

B_PLT_E

numeric, platelet cell count in exp 9/l.

B_WBC_E

numeric, white blood cell count in exp 9/l.

BILI_S

numeric, total bilirubin in µmol/l.

BILID_S

numeric, direct bilirubin in µmol/l.

CA_S

numeric, calcium in mmol/l.

CHE_S

numeric, cholinesterase in µkat/l.

CHOLF_Q

numeric, TODO.

CHOLF_S

numeric, TODO.

CHOLG_S

numeric, total cholesterine in mmol/l.

CL_S

numeric, chloride in mmol/l.

CRE_S

numeric, creatinine in µmol/l.

CRP_S

numeric, C-reactive protein in mg/l.

CYSC_S

numeric, cystatin C in mg/l.

GGT_S

numeric, gamma-glutamyltransferase in µkat/l.

IL6_S

numeric, interleukin 6 in pg/ml.

INR_C

numeric, international normalized ratio.

NA_S

numeric, sodium in mmol/l.

P_S

numeric, phosphate in mmol/l.

PALB_S

numeric, pre-albumin in g/l.

PROT_S

numeric, total protein in g/l.

PTH_S

numeric, parathyreoid hormone in pmol/l.

VDT_OH_S

numeric, vitamin D 25-OH in ng/l.

attr(eldd, "excluded")

integer(3), number of patients that were excluded because they were younger than 18 years, already had a liver transplantation before or their follow-up information was missing/invalid.

attr(eldd, "period")

POSIXct(2), date and times of sampling period (first and last entry).

Source

Institute of Laboratory Medicine, Clinical Chemistry and Molecular Diagnostics, University Hospital Leipzig, Paul-List-Str. 13-15, D-04103 Leipzig, Germany.

Details

The Ethics Committee at the Leipzig University Faculty of Medicine approved the retrospective usage of the data for our study (reference number: 039/14ff).