End-stage Liver Disease Reference

A dataset of reference limits of the University Hospital Leipzig for laboratory measurements used in eldd.

eldr

Format

A data.frame with 36 rows and 8 columns. Each row describes a laboratory measurement/method.

Code

character, name.

Unit

character, unit.

LongDescription

character, full name.

ShortDescription

character, abbreviation.

LowerLimit

numeric, lower limit.

UpperLimit

numeric, upper limit.

AgeDays

integer, the method is only valid for patients older than AgeDays days.

Sex

factor, the method is only valid for "male" or "female" patients. Or the sex doesn't matter ("both").

Source

Institute of Laboratory Medicine, Clinical Chemistry and Molecular Diagnostics, University Hospital Leipzig, Paul-List-Str. 13-15, D-04103 Leipzig, Germany.