Calculate MELD scores

Calculate MELD, MELD-Na and MELD-Plus scores.

meld(
  creatinine,
  bilirubin,
  inr,
  dialysis = FALSE,
  cause = "other",
  round = FALSE
)

meld_na(
  creatinine,
  bilirubin,
  inr,
  sodium,
  dialysis = FALSE,
  type = c("kim2008", "unos"),
  cause = "other",
  round = FALSE
)

meld_plus7(
  creatinine,
  bilirubin,
  inr,
  sodium,
  albumin,
  wbc,
  age,
  round = FALSE
)

Arguments

creatinine

numeric, [mg/dl]

bilirubin

numeric, [mg/dl]

inr

numeric

dialysis

logical, had dialysis twice, or 24 hours of CVVHD, within a week prior to the serum creatinine test?

cause

character, cause of cirrhosis. Has to be of the same length as creatinine, bilirubin, inr and dialysis. Just "ethyltoxic", "cholestatic", "unos" and "other" are supported. Use "unos" for the United Network for Organ Sharing definition of the score.

round

logical, round to nearest integer?

sodium

numeric, [mmol/l]

type

character(1), type of implementation

albumin

numeric, [g/dl]

wbc

numeric, [Gpt/l]

age

numeric, [years]

Value

numeric

Details

Laboratory values (creatinine, bilirubin, INR) below 1.0 are set to 1.0 and creatinine above 4.0 is set to 4.0.

In contrast to meld and meldNa, meldPlus7 returns a probability.

References

Michael Malinchoc et al. 2000. "A model to predict poor survival in patients undergoing transjugular intrahepatic portosystemic shunts" Hepatology 31 (4): 864-871. doi: 10.1053/he.2000.5852

United Network for Organ Sharing (UNOS). Policy notice 11/2015: OPTN executive committee actions. Available at: https://optn.transplant.hrsa.gov/media/1575/policynotice_20151101.pdf Accessed: September 11, 2019.

Kim et al. 2008. "Hyponatremia and Mortality among Patients on the Liver-Transplant Waiting List" N Engl J Med 2008; 359: 1018-1026 doi: 10.1056/NEJMoa0801209

Kartoun et al. 2017. "The MELD-Plus: A generalizable prediction risk score in cirrhosis" PloS one, 12 (10): e0186301 doi: 10.1371/journal.pone.0186301

Examples

meld(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, cause = "other")
#> [1] 20.03916

meld_na(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, sodium = 135)
#> [1] 22.53427
meld_na(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, sodium = 135,
        type = "unos")
#> [1] 21.35658
meld_plus7(2.5, 4.1, 1.2, 137, 24, 6.7, 56)
#> [1] 0.001804553